+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-# Rogue Redux is now in version 0.1.1 !
+# Rogue Redux is now in version 0.1.2 !
+
+## [0.1.2] - 2017-02-27
+### Interface
+- Added savegame button
+- Load game from menu is now functionnal
+### GamePlay
+- Player can now save game at will
+- Player can load a save game from main menu
+- New game erase existing saved game
 
 ## [0.1.1] - 2017-02-26
 ### Interface
@@ -7,7 +16,6 @@
 - Log window functionnal, with basic messages
 - Changed some player fu...ing quotes
 - Log window is now minimizable
-
 
 ## [0.1.0] - 2017-02-25
 ### Added